backend/requirements.txt
flask
transformers
torch
Pillow
numpy
protobuf
scikit-learn